api/resume-ai/requirements.txt
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
openai==0.28.0
gunicorn==21.2.0